Wrote Snake controlled using DIPPID
Needs DIPPID device on port 5700 to control the snake. The goal is to collect as many eggs per minute as possible without eating yourself. The snake speeds up the longer it is. Golden Eggs are 3 times as nutritious. A requirements.txt for a Virtual Environment is given, but almost no packages are needed overall.